requirements.txt
langchain-community
langchain-core
python-dotenv
llama-cpp-python>=0.2.0
pandas
fastapi
uvicorn
rapidfuzz